requirements.txt
docopt~=0.6.2
py-postgresql~=1.2.1
PyYAML~=3.12
